+++ src/main/java/com/torrens/musicshop/domain/Instrument.java
@@ -10,8 +10,7 @@ public class Instrument {
 
     @Id
     @GeneratedValue(strategy = GenerationType.AUTO)
-    private Integer id;
-
+    private Long id;
     private String type;
     private String title;
     private String description;
@@ -27,11 +26,11 @@ public class Instrument {
         this.price = price;
     }
 
-    public Integer getId() {
+    public Long getId() {
         return id;
     }
 
-    public void setId(Integer id) {
+    public void setId(Long id) {
         this.id = id;
     }
